services/account_setup/app/requirements.txt

fastapi
uvicorn
psycopg2-binary
python-dotenv
SQLAlchemy
email-validator
httpx
